add script for csr testing counts

+++ LeSpMV/include/spmv_testroutine.h
@@ -8,7 +8,7 @@ template <typename IndexType, typename ValueType>
 int test_coo_matrix_kernels(const CSR_Matrix<IndexType,ValueType> &csr_ref, int kernel_tag);
 
 template <typename IndexType, typename ValueType>
-int test_csr_matrix_kernels(const CSR_Matrix<IndexType,ValueType> &csr_ref, int kernel_tag, int schedule_mod);
+double test_csr_matrix_kernels(const CSR_Matrix<IndexType,ValueType> &csr_ref, int kernel_tag, int schedule_mod);
 
 template <typename IndexType, typename ValueType>
 int test_bsr_matrix_kernels(const CSR_Matrix<IndexType,ValueType> &csr_ref, int kernel_tag, int schedule_mod);
